Reduce and remove complexity
